+++ CHANGELOG.md
@@ -6,9 +6,14 @@
 
 - Added playsets, a new way to customize your game's rules. Hide moves, oracles, or assets you don't want to use.
 - Added default playsets for Starforged + Sundered Isles for automatically showing & hiding the assets that are relevant to each.
+- Added an option to automatically roll the cursed die to the playset settings.
+- Added cursed die roll to oracle roll notifications and note embeds. 
 
 ### Changes
+
+- Oracles with a cursed version now easily allow you to switch between them.
 
 ### Bug Fixes
 
 - Fixed an issue where replacement moves (like `Repair` from Sundered Isles) were not properly overwriting their base moves.
+- Fixed the Exploration Move Category appearing twice.